asteroid poly and collision tweaks - also teleporting
reworked poly draw so that it doesn't produce leftover tails on polygon
for asteroid - still not massively happy with it. Have added a 5% buffer
for collisions to compensate. Also added a teleport feature that moves
the ship to a random x,y - once number of asteroids is high enough there
is a chance the player can be "boxed in", this gives a risky last minute
option to teleport out - but could result in player teleporting into an
asteroid